requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
requests==2.31.0
bleak==0.21.1
asyncio-mqtt==0.16.1
matplotlib==3.8.4
pyopenssl==24.0.0
sqlcipher3==0.5.2

# Security Dependencies (R11 Implementation)
PyJWT==2.8.0
cryptography==41.0.7
python-dotenv==1.0.0
pywebpush==1.14.0
twilio==8.10.0